Move tests
Move the tests specific to the `LaunchProfileProjectProperties` type to their own file. Tests specific to the `LaunchProfileProjectPropertiesProvider` type remain where they are.
